requirements.txt
Flask==2.0.1
Flask-MySQLdb==0.2.0
Flask-Bcrypt==0.7.1
boto3==1.18.64
SQLAlchemy==1.4.22
mysqlclient==2.1.0
